requirements.txt
aiofiles==23.2.1
anyio==4.2.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
m3u8==4.0.0
nest-asyncio==1.6.0
orjson==3.9.14
sniffio==1.3.0
tqdm==4.66.2
twitter-api-client==0.10.17
uvloop==0.19.0
websockets==12.0
